refactor (ui): decouple UI components from TimetableUiState
