+++ app/src/main/java/com/nelsonxilv/gstoutimetable/presentation/components/TimetableAppBar.kt
@@ -25,27 +25,20 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.nelsonxilv.gstoutimetable.R
 import com.nelsonxilv.gstoutimetable.domain.entity.Group
-import com.nelsonxilv.gstoutimetable.presentation.screen.TimetableUiState
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun TimetableAppBar(
     scrollBehavior: TopAppBarScrollBehavior,
-    timetableUiState: TimetableUiState,
+    titleAppBar: String,
     modifier: Modifier = Modifier
 ) {
-
-    val topAppBarText = if (timetableUiState is TimetableUiState.Success) {
-        timetableUiState.currentGroup
-    } else {
-        stringResource(id = R.string.app_name)
-    }
 
     TopAppBar(
         scrollBehavior = scrollBehavior,
         title = {
             Text(
-                text = topAppBarText,
+                text = titleAppBar,
                 style = MaterialTheme.typography.headlineSmall,
             )
         },
